+++ portal-impl/test/integration/com/liferay/portal/service/http/UserServiceHttpTest.java
@@ -78,8 +78,8 @@ public class UserServiceHttpTest {
 		String firstName = "UserServiceHttpTest";
 		String middleName = StringPool.BLANK;
 		String lastName = "UserServiceHttpTest";
-		int prefixId = 0;
-		int suffixId = 0;
+		long prefixId = 0;
+		long suffixId = 0;
 		boolean male = true;
 		int birthdayMonth = Calendar.JANUARY;
 		int birthdayDay = 1;
