+++ portal-service/src/com/liferay/portal/security/ldap/LDAPUser.java
@@ -349,7 +349,7 @@ public class LDAPUser {
 		_portraitBytes = portraitBytes;
 	}
 
-	public void setPrefixId(int prefixId) {
+	public void setPrefixId(long prefixId) {
 		_contact.setPrefixId(prefixId);
 	}
 
@@ -389,7 +389,7 @@ public class LDAPUser {
 		_user.setStatus(status);
 	}
 
-	public void setSuffixId(int suffixId) {
+	public void setSuffixId(long suffixId) {
 		_contact.setSuffixId(suffixId);
 	}
